requirements.txt
datawire_mdk
docopt
flask
pyyaml
requests
